Added decimal percentages and more ParsedInput adjustments
-Adjusted PercentParserComponent's regex to allow a max of 3 decimals in percentages
-Changed ParsedInput percentage from int to double
-Added an InputDurationTypes enum and changed ParsedInput duration types from a string to this new enum type
-Made the casing of ParsedInput field names consistent with the rest of the project
-Revised percentage parser tests to account for decimal percentages

+++ TRBot/TRBot.VirtualControllers/Controllers/DummyController.cs
@@ -128,7 +128,7 @@ namespace TRBot.VirtualControllers
             InputReleasedEvent?.Invoke(inputName);
         }
 
-        public void PressAxis(in int axis, in double minAxisVal, in double maxAxisVal, in int percent)
+        public void PressAxis(in int axis, in double minAxisVal, in double maxAxisVal, in double percent)
         {
             AxisPressedEvent?.Invoke(axis, percent);
         }
@@ -168,7 +168,7 @@ namespace TRBot.VirtualControllers
             return InputTracker.GetButtonState(buttonVal);
         }
 
-        public int GetAxisState(in int axisVal)
+        public double GetAxisState(in int axisVal)
         {
             return InputTracker.GetAxisState(axisVal);
         }
